+++ oauth2/src/main/scala/org/broadinstitute/dsde/workbench/oauth2/OpenIDConnectConfiguration.scala
@@ -49,6 +49,7 @@ trait OpenIDConnectConfiguration {
 
 object OpenIDConnectConfiguration {
   private val oidcMetadataUrlSuffix = ".well-known/openid-configuration"
+  private[oauth2] val policyParam = "p"
 
   def apply[F[_]: Async](authorityEndpoint: String,
                          oidcClientId: ClientId,